GH-935 Move API calls to their own JS file
Makes it easier to manage and document API calls.
